+++ source/file_input.cpp
@@ -18,10 +18,10 @@ const std::array<Scene::rotation_t, 3> Scene::rotation_handler = {
 
 
 LightSource* Scene::read_source(nlohmann::json& descr) {
-  std::array<float, 3> raw_pos = descr.at("position");
+  std::array<double, 3> raw_pos = descr.at("position");
   std::array<float, 3> intensity = descr.at("intensity");
 
-  Eigen::Vector3f pos(raw_pos[0], raw_pos[1], raw_pos[2]);
+  Eigen::Vector3d pos(raw_pos[0], raw_pos[1], raw_pos[2]);
 
   LightSource* source = new LightSource(pos, LightIntensity(intensity));
 
@@ -57,12 +57,12 @@ BaseObject* Scene::read_sphere(nlohmann::json& descr) {
 
   BaseObject* obj = new Sphere(col, ind);
 
-  float rad = descr.at("radius");
+  double rad = descr.at("radius");
   if (rad != 1.0) {
     obj = Transformation::Scaling(obj, rad, rad, rad);
   }
 
-  std::array<float, 3> pos = descr.at("position");
+  std::array<double, 3> pos = descr.at("position");
   if (pos[0] != 0 or pos[1] != 0 or pos[2] != 0) {
     obj = Transformation::Translation(obj, pos[0], pos[1], pos[2]);
   }
@@ -74,12 +74,12 @@ BaseObject* Scene::read_half_space(nlohmann::json& descr) {
   ColData col = read_col_data(descr.at("color"));
   float ind = descr.at("index");
 
-  std::array<float, 3> normal_arr = descr.at("normal");
-  Eigen::Vector4f normal(normal_arr[0], normal_arr[1], normal_arr[2], 0);
+  std::array<double, 3> normal_arr = descr.at("normal");
+  Eigen::Vector4d normal(normal_arr[0], normal_arr[1], normal_arr[2], 0);
 
   BaseObject* obj = new HalfSpace(col, ind, normal);
 
-  std::array<float, 3> pos = descr.at("position");
+  std::array<double, 3> pos = descr.at("position");
   if (pos[0] != 0 or pos[1] != 0 or pos[2] != 0) {
     obj = Transformation::Translation(obj, pos[0], pos[1], pos[2]);
   }
@@ -88,7 +88,7 @@ BaseObject* Scene::read_half_space(nlohmann::json& descr) {
 }
 
 BaseObject* Scene::read_scaling(nlohmann::json& descr) {
-  std::array<float, 3> fac = descr.at("factors");
+  std::array<double, 3> fac = descr.at("factors");
 
   auto j = descr["subject"].begin();
   BaseObject* subj = action_handler.at(j.key())(j.value());
@@ -99,7 +99,7 @@ BaseObject* Scene::read_scaling(nlohmann::json& descr) {
 }
 
 BaseObject* Scene::read_rotation(nlohmann::json& descr) {
-  float ang = descr["angle"].get<float>() / 180.0 * M_PI;
+  double ang = descr["angle"].get<double>() / 180.0 * M_PI;
   unsigned dir = descr["direction"];
 
   auto j = descr["subject"].begin();
@@ -111,7 +111,7 @@ BaseObject* Scene::read_rotation(nlohmann::json& descr) {
 }
 
 BaseObject* Scene::read_translation(nlohmann::json& descr) {
-  std::array<float, 3> fac = descr.at("factors");
+  std::array<double, 3> fac = descr.at("factors");
 
   auto j = descr["subject"].begin();
   BaseObject* subj = action_handler.at(j.key())(j.value());
@@ -178,7 +178,7 @@ Scene Scene::read_parameters(std::istream& input) {
   RootObject* root = new RootObject(objects);
 
   return Scene(dpi, dim[0], dim[1], 
-               Eigen::Vector4f(pos[0], pos[1], pos[2], 1), Eigen::Vector4f(obs[0], obs[1], obs[2], 1),
+               Eigen::Vector4d(pos[0], pos[1], pos[2], 1), Eigen::Vector4d(obs[0], obs[1], obs[2], 1),
                LightIntensity(amb), index, recursion, sources, root);
 }
